Refactor `LionWebClient` to `LionWebBulkClient` and clean up deprecated classes/interfaces
- Replaced references to `LionWebClient` with `LionWebBulkClient` across the codebase.
- Removed unused and obsolete classes/interfaces (`DeltaErrors`, `SerializationDecorator`, `UnexistingNodeException`, etc.).
- Introduced `BulkRequestFailureException` to handle request failures specific to bulk APIs.
- Updated documentation for improved clarity and consistency.
