+++ client/src/functionalTest/java/io/lionweb/client/LionWebClientDBAdminApiFunctionalTest.java
@@ -22,8 +22,8 @@ public class LionWebClientDBAdminApiFunctionalTest extends AbstractClientFunctio
 
   @Test
   public void repositoriesCRUD() throws IOException {
-    LionWebClient client =
-        new LionWebClient(LionWebVersion.v2024_1, "localhost", getServerPort(), "default");
+    LionWebBulkClient client =
+        new LionWebBulkClient(LionWebVersion.v2024_1, "localhost", getServerPort(), "default");
 
     // Initially we should have one repository
     RepositoryConfiguration defaultRepository =
