+++ client/src/main/java/io/lionweb/client/BulkRequestFailureException.java
@@ -1,11 +1,16 @@
 package io.lionweb.client;
 
-public class RequestFailureException extends RuntimeException {
+/**
+ * This exception is thrown to signal that a request has failed. It includes relevant details about
+ * the failed request, such as the URL, response code, and the response body to provide context for
+ * the failure.
+ */
+public class BulkRequestFailureException extends RuntimeException {
   private final String url;
   private final int responseCode;
   private final String responseBody;
 
-  public RequestFailureException(String url, int responseCode, String responseBody) {
+  public BulkRequestFailureException(String url, int responseCode, String responseBody) {
     super("Request to " + url + " failed with code " + responseCode + ": " + responseBody);
     this.url = url;
     this.responseCode = responseCode;
