--- client/src/main/java/io/lionweb/client/DeltaErrors.java
@@ -1,12 +1,0 @@
-package io.lionweb.client;
-
-public enum DeltaErrors {
-  invalidParticipation,
-  nodeAlreadyExists,
-  unknownNode,
-  unknownIndex,
-  indexNodeMismatch,
-  moveWithoutParent,
-  invalidMove,
-  undefinedReferenceTarget
-}
