+++ client/src/main/java/io/lionweb/client/Protocol.java
@@ -1,5 +1,6 @@
 package io.lionweb.client;
 
+/** Enum representing communication protocols for commonly used protocols HTTP and HTTPS. */
 public enum Protocol {
   HTTP("http"),
   HTTPS("https");
