--- client/src/main/java/io/lionweb/client/SerializationDecorator.java
@@ -1,7 +1,0 @@
-package io.lionweb.client;
-
-import io.lionweb.serialization.JsonSerialization;
-
-public interface SerializationDecorator {
-  void apply(JsonSerialization serialization);
-}
