--- client/src/main/java/io/lionweb/client/UnexistingNodeException.java
@@ -1,19 +1,0 @@
-package io.lionweb.client;
-
-public class UnexistingNodeException extends RuntimeException {
-  private final String nodeID;
-
-  public UnexistingNodeException(String nodeID) {
-    super("Unexisting node " + nodeID);
-    this.nodeID = nodeID;
-  }
-
-  public UnexistingNodeException(String nodeID, Throwable cause) {
-    super("Unexisting node " + nodeID, cause);
-    this.nodeID = nodeID;
-  }
-
-  public String getNodeID() {
-    return nodeID;
-  }
-}
