+++ client/src/main/java/io/lionweb/client/impl/BulkAPIsLionWebClientImplHelper.java
@@ -3,7 +3,7 @@ package io.lionweb.client.impl;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import io.lionweb.client.RequestFailureException;
+import io.lionweb.client.BulkRequestFailureException;
 import io.lionweb.client.api.RepositoryVersionToken;
 import io.lionweb.serialization.LowLevelJsonSerialization;
 import io.lionweb.serialization.data.SerializationChunk;
@@ -40,7 +40,7 @@ abstract class BulkAPIsLionWebClientImplHelper extends LionWebClientImplHelper {
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           return responseData.get("ids").getAsJsonArray().asList().stream()
@@ -60,7 +60,7 @@ abstract class BulkAPIsLionWebClientImplHelper extends LionWebClientImplHelper {
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           SerializationChunk serializationBlock =
