+++ client/src/main/java/io/lionweb/client/impl/ClientForBulkAPIs.java
@@ -3,7 +3,7 @@ package io.lionweb.client.impl;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.lionweb.LionWebVersion;
-import io.lionweb.client.RequestFailureException;
+import io.lionweb.client.BulkRequestFailureException;
 import io.lionweb.client.api.BulkAPIClient;
 import io.lionweb.client.api.JSONLevelBulkAPIClient;
 import io.lionweb.client.api.RepositoryVersionToken;
@@ -64,7 +64,7 @@ public class ClientForBulkAPIs extends BulkAPIsLionWebClientImplHelper implement
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           return conf.getJsonSerialization().deserializeToNodes(responseData.get("chunk"));
@@ -108,7 +108,7 @@ public class ClientForBulkAPIs extends BulkAPIsLionWebClientImplHelper implement
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           List<Node> allNodes =
