+++ client/src/main/java/io/lionweb/client/impl/ClientForHistoryAPIs.java
@@ -2,7 +2,7 @@ package io.lionweb.client.impl;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import io.lionweb.client.RequestFailureException;
+import io.lionweb.client.BulkRequestFailureException;
 import io.lionweb.client.api.HistoryAPIClient;
 import io.lionweb.client.api.RepositoryVersionToken;
 import io.lionweb.model.Node;
@@ -35,7 +35,7 @@ public class ClientForHistoryAPIs extends LionWebClientImplHelper implements His
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           return conf.getJsonSerialization().deserializeToNodes(responseData.get("chunk"));
@@ -71,7 +71,7 @@ public class ClientForHistoryAPIs extends LionWebClientImplHelper implements His
           JsonObject responseData = JsonParser.parseString(responseBody).getAsJsonObject();
           boolean success = responseData.get("success").getAsBoolean();
           if (!success) {
-            throw new RequestFailureException(
+            throw new BulkRequestFailureException(
                 request.url().toString(), response.code(), responseBody);
           }
           List<Node> allNodes =
